api/pyproject.toml
[tool.poetry]
name = "better-budget"
version = "0.1.0"
description = "A budget that suits my specific use case."
authors = ["Jack de la Motte <jack.delamotte13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
Flask = "^2.3.2"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
flask-cors = "^4.0.0"
plaid-python = "^16.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
